+++ src/App.cpp
@@ -115,6 +115,8 @@ void App::on_time_reset()
 
 void App::update()
 {
+    _project.history.start_new_commands_group(); // All commands done in one frame are grouped together, and will be done / undone at once.
+
     // First frame the exe is open
     // Since the construction of an App might be in two steps (constructor, and then deserialization)
     // we do our actual construction logic here, to make sure it is done once and only once.
@@ -600,10 +602,8 @@ void App::imgui_menus()
 void App::reset_cameras()
 {
     auto executor = command_executor();
-    executor.wrap_in_commands_group([&]() {
-        _project.camera_2D_manager.reset_camera(executor);
-        _project.camera_3D_manager.reset_camera(executor);
-    });
+    _project.camera_2D_manager.reset_camera(executor);
+    _project.camera_3D_manager.reset_camera(executor);
 }
 
 void App::check_inputs()