+++ src/CommandCore/CommandExecutor_TopLevel.cpp
@@ -3,13 +3,18 @@
 
 namespace Lab {
 
-void CommandExecutor_TopLevel::execute(const Command& command, CommandExecutionContext_Ref const& ctx) const
+void CommandExecutor_TopLevel::execute(const Command& command, CommandExecutionContext_Ref const& ctx, bool store_in_history) const
 {
-    const auto reversible = try_make_reversible(command, _make_reversible_commands_context); // Must be before the execution of the command because we need to retrieve the state of the app before execution to create the reversible command
-    _sub_executor.execute(command, ctx);
-    if (reversible)
+    if (store_in_history)
     {
-        _history.get().push(*reversible, ReversibleCommandMerger_Ref{});
+        auto const reversible = try_make_reversible(command, _make_reversible_commands_context); // Must be before the execution of the command because we need to retrieve the state of the app before execution to create the reversible command
+        _sub_executor.execute(command, ctx);
+        if (reversible)
+            _history.get().push(*reversible, ReversibleCommandMerger_Ref{});
+    }
+    else
+    {
+        _sub_executor.execute(command, ctx);
     }
 }