+++ src/Commands/Command_SetVariableDefaultValue.h
@@ -3,6 +3,7 @@
 #include <stringify/stringify.hpp>
 #include "CommandCore/CommandExecutionContext_Ref.h"
 #include "CommandCore/MakeReversibleCommandContext_Ref.h"
+#include "CommandCore/var_ref_to_string.h"
 
 namespace Lab {
 
@@ -11,7 +12,7 @@ struct ReversibleCommand_SetVariableDefaultValue;
 
 template<typename T>
 struct Command_SetVariableDefaultValue {
-    Cool::SharedVariableStrongRef<T> var_ref;
+    Cool::SharedVariableStrongRef<T> var_ref{};
     T                                default_value{};
 
     void execute(CommandExecutionContext_Ref const&) const
@@ -21,65 +22,67 @@ struct Command_SetVariableDefaultValue {
 
     auto to_string() const -> std::string
     {
-        return fmt::format("Set {}'s default value to {}", var_ref.id(), Cool::stringify(default_value));
+        return fmt::format("Set {}'s default value to {}", Lab::to_string(var_ref), Cool::stringify(default_value));
     }
 
     auto make_reversible(MakeReversibleCommandContext_Ref const&) const
         -> ReversibleCommand_SetVariableDefaultValue<T>
     {
         return ReversibleCommand_SetVariableDefaultValue<T>{
-            .forward_command   = *this,
+            .fwd               = *this,
             .old_default_value = var_ref.variable->default_value(),
         };
+    }
+
+private:
+    // Serialization
+    friend class cereal::access;
+    template<class Archive>
+    void serialize(Archive& archive)
+    {
+        archive(
+            cereal::make_nvp("Variable ref", var_ref),
+            cereal::make_nvp("Default value", default_value)
+        );
     }
 };
 
 template<typename T>
 struct ReversibleCommand_SetVariableDefaultValue {
-    Command_SetVariableDefaultValue<T> forward_command{};
+    Command_SetVariableDefaultValue<T> fwd{};
     T                                  old_default_value{};
 
     void execute(CommandExecutionContext_Ref const& ctx) const
     {
-        forward_command.execute(ctx);
+        fwd.execute(ctx);
     }
 
     void revert(CommandExecutionContext_Ref const&) const
     {
-        forward_command.var_ref.variable->default_value() = old_default_value;
+        fwd.var_ref.variable->default_value() = old_default_value;
     }
 
     auto to_string() const -> std::string
     {
-        return fmt::format("Set {}'s default value from {} to {}", forward_command.var_ref.id(), Cool::stringify(old_default_value), Cool::stringify(forward_command.default_value));
+        return fmt::format("Set {}'s default value from {} to {}", Lab::to_string(fwd.var_ref), Cool::stringify(old_default_value), Cool::stringify(fwd.default_value));
     }
 
     auto merge(ReversibleCommand_SetVariableDefaultValue<T> const&) const -> std::optional<ReversibleCommand_SetVariableDefaultValue<T>>
     {
         return std::nullopt;
     };
+
+private:
+    // Serialization
+    friend class cereal::access;
+    template<class Archive>
+    void serialize(Archive& archive)
+    {
+        archive(
+            cereal::make_nvp("Forward", fwd),
+            cereal::make_nvp("Old default value", old_default_value)
+        );
+    }
 };
 
 } // namespace Lab
-
-namespace cereal {
-
-template<class Archive, typename T>
-void serialize(Archive& archive, Lab::Command_SetVariableDefaultValue<T>& command)
-{
-    archive(
-        cereal::make_nvp("Variable ref", command.var_ref),
-        cereal::make_nvp("Default value", command.default_value)
-    );
-}
-
-template<class Archive, typename T>
-void serialize(Archive& archive, Lab::ReversibleCommand_SetVariableDefaultValue<T>& command)
-{
-    archive(
-        cereal::make_nvp("Forward", command.forward_command),
-        cereal::make_nvp("Old default value", command.old_default_value)
-    );
-}
-
-} // namespace cereal