+++ src/ModulesGraph/ModulesGraph.h
@@ -2,6 +2,7 @@
 #include "Cool/Nodes/Editor.h"
 #include "Cool/OSC/OSCChannel.h"
 #include "Cool/View/GizmoManager.h"
+#include "DirtyFlags.h"
 #include "Module_Compositing/Module_Compositing.h"
 #include "Module_Particles/Module_Particles.h"
 #include "Nodes/NodesConfig.h"
@@ -50,9 +51,11 @@ public:
     void request_rerender_all();
 
     [[nodiscard]] auto is_empty() const -> bool { return _nodes_editor.is_empty(); }
+    [[nodiscard]] auto graph() const -> Cool::NodesGraph const& { return _nodes_editor.graph(); }
     [[nodiscard]] auto graph() -> Cool::NodesGraph& { return _nodes_editor.graph(); }
-    [[nodiscard]] auto regenerate_code_flag() -> Cool::DirtyFlag const& { return _regenerate_code_flag; }
-    [[nodiscard]] auto rerender_all_flag() -> Cool::DirtyFlag const& { return _rerender_all_flag; }
+    [[nodiscard]] auto regenerate_code_flag() const -> Cool::DirtyFlag const& { return _dirty_flags.regenerate_code; }
+    [[nodiscard]] auto rerender_all_flag() const -> Cool::DirtyFlag const& { return _dirty_flags.rerender; }
+    [[nodiscard]] auto dirty_flags() const -> DirtyFlags const& { return _dirty_flags; }
     [[nodiscard]] auto nodes_config(Ui_Ref, Cool::AudioManager&, Cool::NodesLibrary const&) const -> NodesConfig;
     void               debug_show_nodes_and_links_registries_windows(Ui_Ref ui) const;
     /// Function called once on every frame where the time has changed.
@@ -70,6 +73,18 @@ public:
     void imgui_windows(Ui_Ref, Cool::AudioManager&, Cool::NodesLibrary const&) const;
     void submit_gizmos(Cool::GizmoManager&, CommandExecutor const&, Cool::Camera2D const&);
 
+    //----
+    // These functions are mostly here so that Commands can do their job easily
+    //----
+    auto get_main_node_id() const -> Cool::NodeId const&;
+    void set_main_node_id(Cool::NodeId const& id);
+    void add_node(Cool::NodeId const&, Node const&);
+    void add_link(Cool::LinkId const&, Cool::Link const&);
+    void remove_node(Cool::NodeId const&);
+    void remove_link(Cool::LinkId const&);
+    auto try_get_node(Cool::NodeId const&) const -> Node const*;
+    void set_node(Cool::NodeId const&, Node const&);
+
 private:
     void create_and_compile_all_modules(Cool::NodesGraph const&, Cool::NodeId const& root_node_id, Cool::NodesLibrary const&);
     void render_one_module(Module&, Cool::RenderTarget&, SystemValues const&);
@@ -79,9 +94,7 @@ private:
 private:
     mutable Cool::NodesEditor _nodes_editor{};
     mutable Cool::NodeId      _main_node_id{}; // TODO(Modules) Rename as _root_node_id? Or _output_node_id?
-    mutable Cool::NodeId      _node_we_might_want_to_restore_as_main_node_id{};
-    Cool::DirtyFlag           _regenerate_code_flag{}; // TODO(Modules) Rename as graph_has_changed_flag
-    Cool::DirtyFlag           _rerender_all_flag{};
+    DirtyFlags                _dirty_flags{};
 
     mutable Module_Compositing                     _compositing_module{};
     std::vector<std::unique_ptr<ModulesGraphNode>> _particles_module_nodes{}; // TODO(Particles) No need for the unique_ptr (in theory)
@@ -93,12 +106,11 @@ private:
     void serialize(Archive& archive)
     {
         archive(
-            cereal::make_nvp("Compositing Module", _compositing_module),
-            cereal::make_nvp("Particles Module", _particles_module_nodes),
-            cereal::make_nvp("Dirty Flag: Regenerate Code", _regenerate_code_flag),
-            cereal::make_nvp("Dirty Flag: Rerender all", _rerender_all_flag),
-            cereal::make_nvp("Node Editor", _nodes_editor),
-            cereal::make_nvp("Main Node ID", _main_node_id)
+            cereal::make_nvp("Compositing module", _compositing_module),
+            cereal::make_nvp("Particles module", _particles_module_nodes),
+            cereal::make_nvp("Dirty flags", _dirty_flags),
+            cereal::make_nvp("Node editor", _nodes_editor),
+            cereal::make_nvp("Main node ID", _main_node_id)
         );
     }
 };