+++ src/Nodes/NodesConfig.cpp
@@ -1,11 +1,19 @@
 #include "NodesConfig.h"
-#include <Commands/Command_Group.h>
+#include <Commands/Command_AddLink.h>
+#include <Commands/Command_AddNode.h>
+#include <Commands/Command_ChangeNodeDefinition.h>
+#include <Commands/Command_RemoveLink.h>
+#include <Commands/Command_RemoveNode.h>
+#include <Commands/Command_SetMainNodeId.h>
 #include <Commands/Command_SetVariable.h>
+#include <Commands/Command_SetVariableDefaultMetadata.h>
 #include <Commands/Command_SetVariableDefaultValue.h>
+#include <Commands/Command_SetVariableMetadata.h>
 #include <Cool/Dependencies/always_requires_shader_code_generation.h>
 #include <Cool/Nodes/ed.h>
 #include <Nodes/NodesClipboard.h>
 #include <algorithm>
+#include <reg/src/internal/generate_uuid.hpp>
 #include <string>
 #include "CommandCore/make_command.h"
 #include "Cool/Audio/AudioManager.h"
@@ -65,6 +73,24 @@ static auto make_command_set_variable_default_value(Cool::AnyVariableData const&
         .default_value = std::get<Cool::VariableData<T>>(var_data).value,
     };
 }
+// TODO(Settings) Remove
+template<typename T>
+static auto make_command_set_variable_metadata(Cool::AnyVariableData const& var_data, Cool::SharedVariable<T> const& var) -> Command_SetVariableMetadata<T>
+{
+    return Command_SetVariableMetadata<T>{
+        .var_ref  = var.get_ref(),
+        .metadata = std::get<Cool::VariableData<T>>(var_data).metadata,
+    };
+}
+// TODO(Settings) Remove
+template<typename T>
+static auto make_command_set_variable_default_metadata(Cool::AnyVariableData const& var_data, Cool::SharedVariable<T> const& var) -> Command_SetVariableDefaultMetadata<T>
+{
+    return Command_SetVariableDefaultMetadata<T>{
+        .var_ref          = var.get_ref(),
+        .default_metadata = std::get<Cool::VariableData<T>>(var_data).metadata,
+    };
+}
 
 // TODO(Settings) Remove
 static void apply_settings_to_inputs(
@@ -74,14 +100,15 @@ static void apply_settings_to_inputs(
     CommandExecutor const&                command_executor
 )
 {
-    auto command = Command_Group{};
     try
     {
         for (size_t i = 0; i < inputs.size(); ++i)
         {
             std::visit([&](auto&& input) {
-                command.commands.push_back(make_command(make_command_set_variable(settings.at(i), input)));
-                command.commands.push_back(make_command(make_command_set_variable_default_value(settings.at(i), input)));
+                command_executor.execute(make_command_set_variable(settings.at(i), input));
+                command_executor.execute(make_command_set_variable_default_value(settings.at(i), input));
+                command_executor.execute(make_command_set_variable_metadata(settings.at(i), input));
+                command_executor.execute(make_command_set_variable_default_metadata(settings.at(i), input));
             },
                        inputs.at(i));
         }
@@ -97,7 +124,6 @@ static void apply_settings_to_inputs(
             )
         );
     }
-    command_executor.execute(command);
 }
 
 // TODO(Settings) Remove
@@ -160,15 +186,6 @@ void NodesConfig::main_node_toggle(Cool::NodeId const& node_id)
         else
             set_main_node_id({}); // Unselect main node
     }
-}
-
-auto NodesConfig::primary_dirty_flag(bool always_requires_shader_code_generation) const -> Cool::DirtyFlag const&
-{
-    return always_requires_shader_code_generation ? _regenerate_code_flag : _rerender_flag;
-}
-auto NodesConfig::secondary_dirty_flag() const -> Cool::DirtyFlag const&
-{
-    return _regenerate_code_flag; // At the moment only used by Gradient variable when we detect that the number of marks has changed. See `set_value()` of Command_SetVariable.h
 }
 
 void NodesConfig::imgui_above_node_pins(Cool::Node& /* abstract_node */, Cool::NodeId const& id)
@@ -291,23 +308,16 @@ auto NodesConfig::pin_color(Cool::Pin const& pin, size_t pin_index, Cool::Node c
 
 void NodesConfig::on_node_created(Cool::Node& /* abstract_node */, Cool::NodeId const& node_id, Cool::Pin const* pin_linked_to_new_node)
 {
-    _node_we_might_want_to_restore_as_main_node_id = {};
-
     // Don't change main node if we are dragging a link backward.
     if (pin_linked_to_new_node && pin_linked_to_new_node->kind() == Cool::PinKind::Input)
         return;
 
-    if (!pin_linked_to_new_node)
-        _node_we_might_want_to_restore_as_main_node_id = _main_node_id; // Keep track, so that if someone has a main node, creates a separate node, and then plugs that node back into the old main node, we will restore it as the main node, even if it itself is plugged into a node.
-    set_main_node_id(node_id, true /*keep_node_we_might_want_to_restore_as_main_node*/);
+    set_main_node_id(node_id);
 }
 
-void NodesConfig::set_main_node_id(Cool::NodeId const& id, bool keep_node_we_might_want_to_restore_as_main_node)
+void NodesConfig::set_main_node_id(Cool::NodeId const& id)
 {
-    if (!keep_node_we_might_want_to_restore_as_main_node)
-        _node_we_might_want_to_restore_as_main_node_id = {};
-    _main_node_id = id;
-    _regenerate_code_flag.set_dirty();
+    _command_executor.execute(Command_SetMainNodeId{id});
 }
 
 void NodesConfig::on_link_created_between_existing_nodes(Cool::Link const& link, Cool::LinkId const&)
@@ -320,11 +330,6 @@ void NodesConfig::on_link_created_between_existing_nodes(Cool::Link const& link,
     auto next_id = graph().find_node_containing_pin(link.to_pin_id);
     if (next_id == _main_node_id)
         return;
-    if (next_id == _node_we_might_want_to_restore_as_main_node_id && !_node_we_might_want_to_restore_as_main_node_id.underlying_uuid().is_nil())
-    {
-        set_main_node_id(next_id);
-        return;
-    }
     auto const* next_node        = graph().try_get_node<Node>(next_id);
     auto        new_main_node_id = Cool::NodeId{};
     while (next_node)
@@ -333,11 +338,6 @@ void NodesConfig::on_link_created_between_existing_nodes(Cool::Link const& link,
         next_id          = graph().find_node_connected_to_output_pin(next_node->output_pins()[0].id());
         if (next_id == _main_node_id)
             return;
-        if (next_id == _node_we_might_want_to_restore_as_main_node_id && !_node_we_might_want_to_restore_as_main_node_id.underlying_uuid().is_nil())
-        {
-            set_main_node_id(next_id);
-            return;
-        }
         next_node = graph().try_get_node<Node>(next_id);
     }
     if (!new_main_node_id.underlying_uuid().is_nil())
@@ -364,7 +364,36 @@ static auto wants_to_store_particles_count(NodeDefinition const& def) -> bool
     return Cool::String::contains(def.name(), "Init Particles Count"); // HACK to force only a node named "Init Particles Count" to display the particles count.
 }
 
-auto NodesConfig::make_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -> Node
+auto NodesConfig::add_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -> Cool::NodeId
+{
+    return add_node(make_node(cat_id));
+}
+
+auto NodesConfig::add_node(Node const& node) -> Cool::NodeId
+{
+    auto const id = Cool::NodeId{reg::internal::generate_uuid()};
+    _command_executor.execute(Command_AddNode{id, node});
+    return id;
+}
+
+auto NodesConfig::add_link(Cool::Link const& link) -> Cool::LinkId
+{
+    auto const id = Cool::LinkId{reg::internal::generate_uuid()};
+    _command_executor.execute(Command_AddLink{id, link});
+    return id;
+}
+
+void NodesConfig::remove_node(Cool::NodeId const& id, Cool::Node const& node)
+{
+    _command_executor.execute(Command_RemoveNode{id, node.downcast<Node>()});
+}
+
+void NodesConfig::remove_link(Cool::LinkId const& id, Cool::Link const& link)
+{
+    _command_executor.execute(Command_RemoveLink{id, link});
+}
+
+static auto make_node(Cool::NodeDefinitionAndCategoryName const& cat_id, Cool::GetNodeCategoryConfig_Ref get_node_category_config, DirtyFlags const& dirty_flags) -> Node
 {
     auto const def = cat_id.def.downcast<NodeDefinition>();
 
@@ -375,7 +404,7 @@ auto NodesConfig::make_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -
         needs_main_pin ? def.signature().arity : 0,
         def.function_inputs().size(),
     };
-    auto const node_category_config = _get_node_category_config(cat_id.category_name);
+    auto const node_category_config = get_node_category_config(cat_id.category_name);
 
     if (wants_to_store_particles_count(def))
         node.particles_count() = std::make_optional<size_t>(1000);
@@ -398,8 +427,8 @@ auto NodesConfig::make_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -
             [&](auto&& value_input_def)
                 -> Cool::AnySharedVariable { return Cool::SharedVariable{
                                                  value_input_def,
-                                                 primary_dirty_flag(Cool::always_requires_shader_code_generation(value_input_def)),
-                                                 secondary_dirty_flag()
+                                                 dirty_flags.primary(Cool::always_requires_shader_code_generation(value_input_def)),
+                                                 dirty_flags.secondary()
                                              }; },
             value_input_def
         ));
@@ -417,6 +446,11 @@ auto NodesConfig::make_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -
         node.output_pins().push_back(Cool::OutputPin{output_index_name});
 
     return node;
+}
+
+auto NodesConfig::make_node(Cool::NodeDefinitionAndCategoryName const& cat_id) -> Node
+{
+    return Lab::make_node(cat_id, _get_node_category_config, _dirty_flags);
 }
 
 struct PotentialLink {
@@ -440,7 +474,7 @@ auto NodesConfig::copy_nodes() const -> std::string
 
         clipboard.nodes.push_back({node.as_pod(), pos});
 
-        _nodes_editor.graph().for_each_link_connected_to_node(abstract_node, [&](Cool::Link const& link, bool is_connected_to_input_pin) {
+        _nodes_editor.graph().for_each_link_connected_to_node(abstract_node, [&](Cool::LinkId const&, Cool::Link const& link, bool is_connected_to_input_pin) {
             size_t index = potential_links.size();
             for (size_t i = 0; i < potential_links.size(); ++i)
             {
@@ -481,7 +515,7 @@ auto NodesConfig::paste_nodes(std::string_view clipboard_string) -> bool
     {
         auto              clipboard = string_to_nodes_clipboard(std::string{clipboard_string});
         std::vector<Node> new_nodes{};
-        // Create all nodes but don't add them to the graph yet, because we need to call update_node_with_new_definition() first, which requires all the links to have been added to the graph. And before adding all the links we must first iterate over all the nodes and update the links with the new pin ids of the nodes.
+        // Create all nodes but don't add them to the graph yet, because we need to call make_sure_node_uses_the_most_up_to_date_version_of_its_definition() first, which requires all the links to have been added to the graph. And before adding all the links we must first iterate over all the nodes and update the links with the new pin ids of the nodes.
         for (auto const& clipboard_node : clipboard.nodes)
         {
             auto node = Node{clipboard_node.node_as_pod.pod_part};
@@ -490,9 +524,9 @@ auto NodesConfig::paste_nodes(std::string_view clipboard_string) -> bool
                 std::visit([&](auto&& value_input) {
                     node.value_inputs().push_back(Cool::SharedVariable{
                         value_input,
-                        primary_dirty_flag(Cool::always_requires_shader_code_generation(value_input)),
-                        {}, // description, it will be set by update_node_with_new_definition()
-                        secondary_dirty_flag()
+                        _dirty_flags.primary(Cool::always_requires_shader_code_generation(value_input)),
+                        {}, // description, it will be set by make_sure_node_uses_the_most_up_to_date_version_of_its_definition()
+                        _dirty_flags.secondary()
                     });
                 },
                            value_input);
@@ -521,19 +555,18 @@ auto NodesConfig::paste_nodes(std::string_view clipboard_string) -> bool
         }
 
         for (auto const& link : clipboard.links)
-            graph().add_link_unchecked(link);
+            add_link(link);
 
         for (size_t i = 0; i < new_nodes.size(); ++i)
         {
             auto const& node_in_clipboard = clipboard.nodes[i];
             auto&       node              = new_nodes[i];
 
-            auto const* node_def = _get_node_definition(node.id_names());
-            if (node_def)
-                // Must be done after all the links have been added to the graph, because it might remove some that are outdated if some pins have been removed and they had a link connected to them.
-                update_node_with_new_definition(node, *node_def, graph()); // Check if the definition has changed (e.g. new inputs) and also finds description of variables if any.
+            // Check if the definition has changed (e.g. new inputs) and also finds description of variables if any.
+            // Must be done after all the links have been added to the graph, because it might remove some that are outdated if some pins have been removed and they had a link connected to them.
+            make_sure_node_uses_the_most_up_to_date_version_of_its_definition(node);
 
-            auto const new_node_id    = graph().add_node(node);
+            auto const new_node_id    = add_node(node);
             auto const new_node_id_ed = Cool::as_ed_id(new_node_id);
 
             ed::SetNodePosition(new_node_id_ed, ImGui::GetMousePos() + node_in_clipboard.position);
@@ -587,7 +620,7 @@ static void keep_values_of_inputs_that_already_existed_and_destroy_unused_ones(
 }
 
 template<typename PinT>
-static void refresh_pins(std::vector<PinT>& new_pins, std::vector<PinT> const& old_pins, std::function<void(Cool::PinId const&)> remove_links)
+static void refresh_pins(std::vector<PinT>& new_pins, std::vector<PinT> const& old_pins, std::function<void(Cool::PinId const&)> const& remove_links)
 {
     // Collect indices of pins that don't match any old pins.
     // Start with all the indices of the new input pins
@@ -625,24 +658,65 @@ static void refresh_pins(std::vector<PinT>& new_pins, std::vector<PinT> const& o
     }
 }
 
-void NodesConfig::update_node_with_new_definition(Cool::Node& abstract_out_node, Cool::NodeDefinition const& definition, Cool::NodesGraph& graph)
+static void refresh_pins(Node& node, Node const& out_node, std::function<void(Cool::PinId const&)> const& remove_links)
 {
-    update_node_with_new_definition(abstract_out_node.downcast<Node>(), definition, graph);
+    refresh_pins(node.input_pins(), out_node.input_pins(), remove_links);
+    refresh_pins(node.output_pins(), out_node.output_pins(), remove_links);
 }
 
-void NodesConfig::update_node_with_new_definition(Node& out_node, Cool::NodeDefinition const& definition, Cool::NodesGraph& graph)
+void NodesConfig::change_node_definition(Cool::NodeId const& id, Cool::Node& node, Cool::NodeDefinition const& def)
 {
-    auto node = make_node({definition, out_node.category_name()});
+    auto node_copy = node.downcast<Node>(); // Don't change the original node just yet, it will be done by the command
+    update_node_with_new_definition(node_copy, def, true /*store_links_deletion_in_history*/);
+    _command_executor.execute(Command_ChangeNodeDefinition{id, std::move(node_copy)});
+}
+
+void NodesConfig::update_node_with_new_definition(Cool::Node& abstract_out_node, Cool::NodeDefinition const& definition)
+{
+    update_node_with_new_definition(abstract_out_node.downcast<Node>(), definition);
+}
+
+static void update_node_with_new_definition(Node& out_node, Cool::NodeDefinition const& definition, Cool::NodesGraph& graph, CommandExecutor const& command_executor, Cool::GetNodeCategoryConfig_Ref get_node_category_config, DirtyFlags const& dirty_flags, bool store_links_deletion_in_history = false)
+{
+    auto node = make_node({definition, out_node.category_name()}, get_node_category_config, dirty_flags);
     node.set_name(out_node.name());
     if (node.particles_count().has_value() && out_node.particles_count().has_value())
         node.set_particles_count(out_node.particles_count());
 
     keep_values_of_inputs_that_already_existed_and_destroy_unused_ones(out_node.value_inputs(), node.value_inputs());
 
-    refresh_pins(node.input_pins(), out_node.input_pins(), [&](Cool::PinId const& pin_id) { graph.remove_link_going_into(pin_id); });
-    refresh_pins(node.output_pins(), out_node.output_pins(), [&](Cool::PinId const& pin_id) { graph.remove_link_coming_from(pin_id); });
+    if (store_links_deletion_in_history)
+    {
+        auto links_to_remove = std::vector<std::pair<Cool::LinkId, Cool::Link>>{};
+        refresh_pins(node, out_node, [&](Cool::PinId const& pin_id) {
+            graph.for_each_link_connected_to_pin(pin_id, [&](Cool::LinkId const& id, Cool::Link const& link, bool) { links_to_remove.emplace_back(id, link); });
+        });
+        for (auto const& [id, link] : links_to_remove)
+            command_executor.execute(Command_RemoveLink{id, link});
+    }
+    else
+    {
+        refresh_pins(node, out_node, [&](Cool::PinId const& pin_id) { graph.remove_link_going_into(pin_id); });
+    }
 
     out_node = std::move(node);
+}
+
+void NodesConfig::update_node_with_new_definition(Node& out_node, Cool::NodeDefinition const& definition, bool store_links_deletion_in_history)
+{
+    Lab::update_node_with_new_definition(out_node, definition, graph(), _command_executor, _get_node_category_config, _dirty_flags, store_links_deletion_in_history);
+}
+
+void make_sure_node_uses_the_most_up_to_date_version_of_its_definition(Node& node, Cool::GetNodeDefinition_Ref<NodeDefinition> get_node_definition, Cool::NodesGraph& graph, CommandExecutor const& command_executor, Cool::GetNodeCategoryConfig_Ref get_node_category_config, DirtyFlags const& dirty_flags)
+{
+    auto const* const node_def = get_node_definition(node.id_names());
+    if (node_def)
+        Lab::update_node_with_new_definition(node, *node_def, graph, command_executor, get_node_category_config, dirty_flags);
+}
+
+void NodesConfig::make_sure_node_uses_the_most_up_to_date_version_of_its_definition(Node& node)
+{
+    Lab::make_sure_node_uses_the_most_up_to_date_version_of_its_definition(node, _get_node_definition, graph(), _command_executor, _get_node_category_config, _dirty_flags);
 }
 
 void NodesConfig::widget_to_rename_node(Cool::Node& abstract_node)